+++ sprite.h
@@ -22,7 +22,7 @@ public:
 	virtual void draw(Context &ctx);
 	virtual void update(Context &ctx);
 
-	template <int C> float final() const {
+	template <int C> double final() const {
 		return std::get<C>(m_home) + std::get<C>(m_relpos);
 	}
 
@@ -34,7 +34,7 @@ protected:
 	const Texture *m_texture;
 	Point m_relpos;
 	Point m_home;
-	GLfloat m_size;
+	GLdouble m_size;
 };
 
 class Yonker : public Sprite {
@@ -45,7 +45,7 @@ public:
 		AMBITION = 2,
 		_EMOTIONS_COUNT
 	};
-	using EmotionVector = std::array<float, _EMOTIONS_COUNT>;
+	using EmotionVector = std::array<double, _EMOTIONS_COUNT>;
 
 	Yonker(const Texture *texture, const Point &home);
